found typo, still throwing an error

+++ saxpy/saxpy.cu
@@ -96,7 +96,7 @@ void saxpyCuda(int N, float alpha, float* xarray, float* yarray, float* resultar
     //
     // CS149 TODO: copy result from GPU back to CPU using cudaMemcpy
     //
-    cudaMemcpy(device_result, resultarray, N * sizeof(int), cudaMemcpyDeviceToHost);
+    cudaMemcpy(device_result, resultarray, N * sizeof(float), cudaMemcpyDeviceToHost);
     // end timing after result has been copied back into host memory
     double endTime = CycleTimer::currentSeconds();
